# api/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Go Demo API with user authentication and comments",
    "title": "Go Demo API",
    "contact": {
      "name": "API Support",
      "email": "alan.e.george86@gmail.com"
    },
    "version": "1.0"
  },
  "basePath": "/",
  "paths": {
    "/comment": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Post a new comment (requires JWT authentication)",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["comments"],
        "summary": "Create a comment",
        "parameters": [
          {
            "description": "Comment content",
            "name": "comment",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/main.CreateCommentRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/main.CommentResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/main.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/main.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/main.ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Delete a comment (requires JWT authentication)",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["comments"],
        "summary": "Delete a comment",
        "parameters": [
          {
            "description": "Comment deletion details",
            "name": "comment",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/main.DeleteCommentRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/main.DeleteCommentSuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/main.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/main.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/main.ErrorResponse"
            }
          }
        }
      }
    },
    "/comments": {
      "get": {
        "description": "List all comments",
        "produces": ["application/json"],
        "tags": ["comments"],
        "summary": "List all comments",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/main.CommentResponse"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/main.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/main.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/main.ErrorResponse"
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "description": "Authenticate with username and password to receive a JWT token",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["authentication"],
        "summary": "User login",
        "parameters": [
          {
            "description": "Login credentials",
            "name": "credentials",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/main.LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/main.LoginResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/main.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/main.ErrorResponse"
            }
          }
        }
      }
    },
    "/user": {
      "post": {
        "description": "Register a new user with name and password",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["users"],
        "summary": "Create a new user",
        "parameters": [
          {
            "description": "User registration details",
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/main.CreateUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/main.UserResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/main.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/main.ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "main.CommentResponse": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "created_at": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "id": {
          "type": "string"
        }
      }
    },
    "main.CreateCommentRequest": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        }
      }
    },
    "main.CreateUserRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "main.DeleteCommentRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "id": {
          "type": "string"
        }
      }
    },
    "main.DeleteCommentSuccessResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "main.ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        }
      }
    },
    "main.LoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "main.LoginResponse": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "user": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            },
            "id": {
              "type": "string"
            }
          }
        }
      }
    },
    "main.UserResponse": {
      "type": "object",
      "properties": {
        "created_at": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "id": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Type \"Bearer\" followed by a space and JWT token.",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}